TR-45-refactor: axios인스턴스와 fetcher함수 이용하도록 변경
